% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPCMlasso-package.R
\docType{data}
\name{tenseness}
\alias{tenseness}
\title{Tenseness data from the Freiburg Complaint Checklist}
\format{A data frame containing data from the Freiburg Complaint Checklist with 1847 observations. 
All items refer to the scale \emph{Tenseness} and are measured on a 5-point Likert scale where low numbers 
correspond to low frequencies or low intensitites of the respective complaint and vice versa. 
\describe{ 
\item{Clammy_hands}{Do you have clammy hands?}
\item{Sweat_attacks}{Do you have sudden attacks of sweating?}
\item{Clumsiness}{Do you notice that you behave clumsy?}
\item{Wavering_hands}{Are your hands wavering frequently, e.g. when lightning a cigarette or when holding a cup?}
\item{Restless_hands}{Do you notice that your hands are restless?}
\item{Restless_feet}{Do you notice that your feet are restless?}
\item{Twitching_eyes}{Do you notice unvoluntary twitching of your eyes?}
\item{Twitching_mouth}{Do you notice unvoluntary twitching of your mouth?}
\item{Gender}{Gender of the person}
\item{Household}{Does the person live alone in a household or together with somebody?}
\item{Income}{Income, categorized to levels from 1 (low income) to 11(high income). For simplicity,
due to the high number of categories income can be treated as a metric variable.}
\item{WestEast}{Is the person from East Germany (former GDR)?}
\item{Abitur}{Does the person have Abitur (A-levels)?}
\item{Age}{Age of the person}
 }}
\source{
ZPID (2013). PsychData of the Leibniz Institute for Psychology Information ZPID. Trier: Center for Research Data in Psychology.

Fahrenberg, J. (2010). Freiburg Complaint Checklist [Freiburger Beschwerdenliste (FBL)]. Goettingen, Hogrefe.
}
\description{
Data from the Freiburg Complaint Checklist. 
The data contain all 8 items corresponding to the scale \emph{Tenseness} for 2042 participants of the 
standardization sample of the Freiburg Complaint Checklist.
}
\examples{

data(tenseness)

}
\keyword{datasets}
